+++ lib/src/widgets/card_expiry_form_field.dart
@@ -1,8 +1,10 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 import 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
 
 /// Form field to edit a credit card expiration date, with validation.
-class CardExpiryFormField extends StatelessWidget {
+class CardExpiryFormField extends StatefulWidget {
   const CardExpiryFormField(
       {Key key,
       this.initialMonth,
@@ -22,52 +24,56 @@ class CardExpiryFormField extends StatelessWidget {
   final void Function(int, int) onChanged;
   final String Function(String) validator;
 
-  static const defaultLabelText = "Expiry Date";
+  static const defaultLabelText = "Expiration Date";
   static const defaultHintText = "MM/YY";
-  static const defaultErrorText = "Invalid expiry date";
+  static const defaultErrorText = "Invalid expiration date";
   static const defaultMonthMask = "##";
   static const defaultYearMask = "##";
 
-  static const defaultDecoration = InputDecoration(
-      border: OutlineInputBorder(),
-      labelText: defaultLabelText,
-      hintText: defaultHintText);
+  static const defaultDecoration =
+      InputDecoration(border: OutlineInputBorder(), labelText: defaultLabelText, hintText: defaultHintText);
   static const defaultTextStyle = TextStyle(color: Colors.black);
 
   @override
-  Widget build(BuildContext context) {
-    var maskFormatter =
-        MaskTextInputFormatter(mask: '$defaultMonthMask/$defaultYearMask');
+  _CardExpiryFormFieldState createState() => _CardExpiryFormFieldState();
+}
 
-    final month =
-        initialMonth?.toString()?.padLeft(defaultMonthMask.length, "0");
-    final year = initialYear
-        ?.toString()
-        ?.substring(initialYear.toString().length - defaultYearMask.length);
+class _CardExpiryFormFieldState extends State<CardExpiryFormField> {
+  final maskFormatter =
+      MaskTextInputFormatter(mask: '${CardExpiryFormField.defaultMonthMask}/${CardExpiryFormField.defaultYearMask}');
+
+  @override
+  Widget build(BuildContext context) {
+    final month = widget.initialMonth?.toString()?.padLeft(CardExpiryFormField.defaultMonthMask.length, "0");
+    final year = widget.initialYear?.toString()?.substring(widget.initialYear.toString().length -
+        min(CardExpiryFormField.defaultYearMask.length, widget.initialYear.toString().length));
     final initial = (month ?? "") + (year ?? "");
+
+    final initialMaskFormatter =
+        MaskTextInputFormatter(mask: '${CardExpiryFormField.defaultMonthMask}/${CardExpiryFormField.defaultYearMask}');
 
     return Container(
       child: TextFormField(
-        validator: validator,
-        initialValue: maskFormatter
-            .formatEditUpdate(
-                TextEditingValue(), TextEditingValue(text: initial))
-            .text,
+        validator: widget.validator,
+        initialValue: initialMaskFormatter.formatEditUpdate(TextEditingValue(), TextEditingValue(text: initial)).text,
         onChanged: (text) {
           final arr = text.split("/");
           final month = int.tryParse(arr[0]);
-          final year = int.tryParse(arr[1]);
-          onChanged(month, year);
+          var year;
+          if (arr.length == 2) {
+            year = int.tryParse(arr[1]);
+          }
+          widget.onChanged(month, year);
         },
         onSaved: (text) {
           final arr = text.split("/");
           final month = int.tryParse(arr[0]);
           final year = int.tryParse(arr[1]);
-          onSaved(month, year);
+          widget.onSaved(month, year);
         },
         inputFormatters: [maskFormatter],
-        style: textStyle,
-        decoration: decoration,
+        style: widget.textStyle,
+        decoration: widget.decoration,
         keyboardType: TextInputType.number,
         textInputAction: TextInputAction.next,
         onFieldSubmitted: (value) => FocusScope.of(context).nextFocus(),